buffer messages before writing to file
Add two properties to the logging class
bufferMessages : The first is a cell array that contains messages to be
written.
bufferCursor_ is an integer that points to next cell to write in
bufferMessages_

During logging, messages are added to buffer. Once it's full, log lines
are written on file.

This new feature have no impact on logging on console.

A new method called setBufferingSize was created to change buffer size, on the model of the
others methods